backend/requirements.txt
# Lexra Backend — Azure Production Requirements
# Heavy ML packages (sentence-transformers, faiss-cpu, ollama) excluded
# to keep Oryx build time under the 4-minute timeout.

# FastAPI and core
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==21.2.0
python-multipart==0.0.12
pydantic==2.8.2
pydantic-settings==2.4.0
aiofiles==24.1.0

# Document processing
pdfplumber==0.11.4
python-docx==1.1.2

# Azure SDK
azure-storage-blob==12.23.1
azure-identity==1.19.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# OpenTelemetry
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0

# Async DB
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
pyodbc==5.2.0

# Utilities
httpx==0.27.2
python-dotenv==1.0.1
structlog==24.4.0
